Move and rename tree builder outside of contextual/
